Update User-facing Copyrights for Wren + CDDL
Updates all copyrights to assert Wren copyright for the portions we've added. I believe the CDDL prevents us from removing FR's own copyrights.

Also, modifies which license we link to since this is licensed under the CDDL, not the ForgeRock commercial license. Links appear in a new window to prevent issues with licenses replacing the content of AM windows during install.